+++ riotfile.py
@@ -1478,7 +1478,7 @@ venv = Venv(
                 Venv(
                     pys=["2.7"],
                     # pytest==4.6 is last to support python 2.7
-                    pkgs={"pytest": ">=4.0,<4.6", "asynctest": "==0.11.1", "msgpack": latest},
+                    pkgs={"pytest": ">=4.0,<4.6", "msgpack": latest},
                 ),
                 Venv(
                     pys=select_pys(min_version="3.5", max_version="3.9"),
@@ -1491,7 +1491,6 @@ venv = Venv(
                             latest,
                         ],
                         "msgpack": latest,
-                        "asynctest": "==0.13.0",
                         "more_itertools": "<8.11.0",
                     },
                 ),
@@ -1505,6 +1504,21 @@ venv = Venv(
                         "msgpack": latest,
                         "asynctest": "==0.13.0",
                         "more_itertools": "<8.11.0",
+                    },
+                ),
+            ],
+        ),
+        Venv(
+            name="asynctest",
+            command="pytest {cmdargs} tests/contrib/asynctest/",
+            venvs=[
+                Venv(
+                    pys=select_pys(min_version="3.5", max_version="3.9"),
+                    pkgs={
+                        "pytest": [
+                            ">=6.0,<7.0",
+                        ],
+                        "asynctest": "==0.13.0",
                     },
                 ),
             ],
